#67 Move each file into a domain folder
