+++ src/util/paging-helpers.ts
@@ -1,7 +1,3 @@
-import { Role } from '@prisma/client';
-
-import { QuizlordContext } from '..';
-
 export function base64Encode(source: string) {
   return Buffer.from(source).toString('base64');
 }
@@ -20,10 +16,4 @@ export interface PagedResult<T> {
     startCursor?: string;
     endCursor?: string;
   };
-}
-
-export function requireUserRole(context: QuizlordContext, role: Role) {
-  if (!context.roles.includes(role)) {
-    throw new Error('You are not authorised to perform this action');
-  }
 }
